ctkBackTrace: bug fix discovered on MacOSX, the method signature was wrong

+++ Libs/Core/ctkBackTrace.cpp
@@ -165,7 +165,7 @@ int ctkBackTracePrivate::trace(void** array, size_t n) const
 #else
 
 // --------------------------------------------------------------------------
-int ctkBackTracePrivate::trace(void** /*array*/, size_t /*n*/) const
+int ctkBackTracePrivate::trace(void** /*array*/, int /*n*/) const
 {
   return 0;
 }
